Try both v4 and v6 socket types
The libroken-getifaddrs fails to retrive the ipv6-address in
solaris-zones but it might connect over ipv6 anyway, and then the kdc
refuses to give a ticket with the cryptic message:

kinit: krb5_get_init_creds: No ENC-TS found

A saner message ends up in the kdc's logfile.

because of a ENXIO when looking for a ipv6-address on the
ipv4-interface, the whole getlifaddrs2 fails and getifaddrs2 is run
instead and it just discovers the ipv4-address.

Signed-off-by: Love Hornquist Astrand <lha@h5l.org>