Update krb5-config.
We update krb5-config to support --all, --deps, and --vendor.  So,
now if you specify --libs, you will only get those libraries that
are required for dynamic linking.  If you are linking statically,
you must provide --deps.  We also allow multiple libraries to be
specified on the command line for applications that use, say, gssapi
and krb5.  And we elide -L and -I args if they have the prefix=/usr
as that's implied.  We also update the --help output to be a little
more verbose.