Fix the db3 locking.
We experienced concurrency issues when using Berkeley DB 3 because
we were not locking the database.  As it turns out, when using the
database, we must either use Berkeley DB's locking functions or
obtain a lock which spans our entire use of the DB as the library
will cache data read from the file which it will then use to do
things like determine where to read from the file.  Failing to lock
the DB for the entire time results in corrupted databases when
multiple writers are in operation.

We use flock(2) to obtain a lock on a separate file descriptor
open(2)ed on the db file to address this issue.  While we are here,
we also reduce the DB3 cache size as Heimdal's use of DB3 opens
and closes the DB a lot and is unlikely to use the cache significantly.