Windows: $(SDKREDISTDIR)\heimdal-version.mak
Create a heimdal-version.mak include file which can be included
in third party installer makefiles to determine which compiler and run
time library type was used to build the merge modules.  It is the
responsibility of third party installer packages to include the Visual
Studio CRT merge modules that are required for execution.

Change-Id: Icec2f47755ef7cc5f7dde2e0426603dece7e4933